✨ 댓글 추가 컴포넌트 분리
